Move IoVec and writev to separate helper crate
